Add React Router and rough project structure
Add React Router v6 integration to the project. Based on the rough
links, set up the page structure for the user service and question
service in React Router.

Added a simple frontend authentication context to the application to
allow for different views when logged in/logged out, as well as provide
distinction between admin users and non-admin users.